# Conflicts:
#	src/main/java/com/t2s/staying/home/T2S/StayingHome/ApplicationConstants.java
#	src/main/java/com/t2s/staying/home/T2S/StayingHome/factory/CommandsFactory.java

+++ src/main/java/com/t2s/staying/home/T2S/StayingHome/tts/FakeTextToSpeechAPI.java
@@ -7,7 +7,7 @@ public class FakeTextToSpeechAPI implements TextToSpeechAPI {
 
     public void play(String pl){}
 
-    public void setVolume(int volume){}
+    public void setVolume(float volume){}
 
     public void setPitch(int pitch){}
