+++ src/main/java/com/t2s/staying/home/T2S/StayingHome/tts/TextToSpeechAPI.java
@@ -8,7 +8,7 @@ package com.t2s.staying.home.T2S.StayingHome.tts;
 public interface TextToSpeechAPI {
 
     void play(String pl);
-    void setVolume(int volume);
+    void setVolume(float volume);
     void setPitch(int pitch);
     void setRate(int rate);
 }
